Make slogdet, linalg.sloget and logdet support metatensors
This PR also adds complex support for logdet, and makes all these
functions support out= and be composite depending on one function. We
also extend the support of `logdet` to complex numbers and improve the
docs of all these functions.

We also use `linalg_lu_factor_ex` in these functions, so we remove the
synchronisation present before.

ghstack-source-id: 9fc60cd0dba05ab32659fd49a66b3d9059454cd6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79742